+++ module-apps/application-desktop/windows/DesktopMainWindow.cpp
@@ -7,10 +7,9 @@
  * @details
  */
 #include <memory>
-#include <functional>
 
-#include "../ApplicationDesktop.hpp"
-#include "../data/LockPhoneData.hpp"
+#include "application-desktop/ApplicationDesktop.hpp"
+#include "application-desktop/data/LockPhoneData.hpp"
 #include "Alignment.hpp"
 #include "Common.hpp"
 #include "BottomBar.hpp"
@@ -19,11 +18,8 @@
 #include "gui/widgets/Image.hpp"
 #include "service-appmgr/ApplicationManager.hpp"
 
-#include "application-call/ApplicationCall.hpp"
-#include "application-call/data/CallSwitchData.hpp"
 #include <UiCommonActions.hpp>
 
-#include <gui/tools/Common.hpp>
 #include "i18/i18.hpp"
 #include "log/log.hpp"
 #include <Span.hpp>
@@ -31,7 +27,6 @@
 #include <application-settings-new/ApplicationSettings.hpp>
 #include <application-settings/ApplicationSettings.hpp>
 #include <cassert>
-#include <service-appmgr/ApplicationManager.hpp>
 #include <time/time_conversion.hpp>
 
 namespace style
@@ -52,7 +47,6 @@ namespace gui
 
     void DesktopMainWindow::buildInterface()
     {
-
         auto ttime = utils::time::Time();
         AppWindow::buildInterface();
 
@@ -100,7 +94,7 @@ namespace gui
         auto app = dynamic_cast<app::ApplicationDesktop *>(application);
         assert(app != nullptr);
 
-        if (app->getScreenLocked()) {
+        if (app->lock.isLocked()) {
             bottomBar->restore();
             bottomBar->setText(BottomBar::Side::CENTER, utils::localize.get("app_desktop_unlock"));
             topBar->setActive(TopBar::Elements::LOCK, true);
@@ -109,8 +103,6 @@ namespace gui
             erase(notifications);
         }
         else {
-            auto app = dynamic_cast<app::ApplicationDesktop *>(application);
-            assert(app);
             bottomBar->setText(BottomBar::Side::CENTER, utils::localize.get("app_desktop_menu"));
             topBar->setActive(TopBar::Elements::LOCK, false);
             if (!fillNotifications(app)) {
@@ -127,22 +119,91 @@ namespace gui
     {
         // update time
         time->setText(topBar->getTimeString());
-
         // check if there was a signal to lock the pone due to inactivity.
         if ((data != nullptr) && (data->getDescription() == "LockPhoneData")) {
             auto app = dynamic_cast<app::ApplicationDesktop *>(application);
             if (!app) {
                 return;
             }
-            app->setScreenLocked(true);
+            app->lock.isLocked();
 
             LockPhoneData *lockData = reinterpret_cast<LockPhoneData *>(data);
             lockTimeoutApplilcation = lockData->getPreviousApplication();
 
             reinterpret_cast<app::ApplicationDesktop *>(application)->setSuspendFlag(true);
         }
-
         setVisibleState();
+    }
+
+    bool DesktopMainWindow::processLongPressEvent(const InputEvent &inputEvent)
+    {
+        app::ApplicationDesktop *app = dynamic_cast<app::ApplicationDesktop *>(application);
+        if (app == nullptr) {
+            LOG_ERROR("not ApplicationDesktop");
+            return AppWindow::onInput(inputEvent);
+        }
+
+        if ((inputEvent.keyCode == KeyCode::KEY_PND) && (!app->lock.isLocked())) {
+            app->lock.lock();
+            setVisibleState();
+            app->refreshWindow(RefreshModes::GUI_REFRESH_FAST);
+            app->setSuspendFlag(true);
+            return true;
+        }
+        else if (inputEvent.keyCode == KeyCode::KEY_RF) {
+            application->switchWindow("PowerOffWindow");
+            return true;
+        }
+        // long press of '0' key is translated to '+'
+        else if (inputEvent.keyCode == KeyCode::KEY_0) {
+            return app::prepareCall(application, "+");
+        }
+        // check if any of the lower inheritance onInput methods catch the event
+        return AppWindow::onInput(inputEvent);
+    }
+
+    bool DesktopMainWindow::processShortPressEventOnUnlocked(const InputEvent &inputEvent)
+    {
+        auto code = translator.handle(inputEvent.key, InputMode({InputMode::phone}).get());
+        if (inputEvent.keyCode == KeyCode::KEY_ENTER) {
+            application->switchWindow(app::window::name::desktop_menu);
+            return true;
+        }
+        // if numeric key was pressed record that key and send it to call application
+        else if (code != 0) {
+            return app::prepareCall(application, std::string(1, static_cast<char>(code)));
+        }
+        // check if any of the lower inheritance onInput methods catch the event
+        return AppWindow::onInput(inputEvent);
+    }
+
+    bool DesktopMainWindow::processShortPressEventOnLocked(const InputEvent &inputEvent)
+    {
+        app::ApplicationDesktop *app = dynamic_cast<app::ApplicationDesktop *>(application);
+        if (app == nullptr) {
+            LOG_ERROR("not ApplicationDesktop");
+            return AppWindow::onInput(inputEvent);
+        }
+        // if enter was pressed
+        if (enter_cache.cached() && inputEvent.is(KeyCode::KEY_PND)) {
+            // if interval between enter and pnd keys is less than time defined for unlocking
+            // display pin lock screen or simply refresh current window to update labels
+            std::unique_ptr<LockPhoneData> data = std::make_unique<LockPhoneData>(LockPhoneData::Request::Pin);
+            // if there was no application on to before closing proceed normally to pin protection
+            if (lockTimeoutApplilcation.empty()) {
+                application->switchWindow(app::window::name::desktop_pin_lock, std::move(data));
+                return true;
+            }
+            else {
+                data->setPrevApplication(lockTimeoutApplilcation);
+                lockTimeoutApplilcation = "";
+                application->switchWindow(app::window::name::desktop_pin_lock, std::move(data));
+                return true;
+            }
+        }
+        enter_cache.storeEnter(inputEvent);
+        // check if any of the lower inheritance onInput methods catch the event
+        return AppWindow::onInput(inputEvent);
     }
 
     bool DesktopMainWindow::onInput(const InputEvent &inputEvent)
@@ -153,97 +214,17 @@ namespace gui
             return AppWindow::onInput(inputEvent);
         }
 
-        auto code = translator.handle(inputEvent.key, InputMode({InputMode::phone}).get());
-
-        // process shortpress
-        if (inputEvent.isShortPress()) {
-            if (app->getScreenLocked()) {
-                // if enter was pressed
-                if (enter_cache.cached() && inputEvent.is(KeyCode::KEY_PND)) {
-                    // if interval between enter and pnd keys is less than time defined for unlocking
-                    // display pin lock screen or simply refresh current window to update labels
-                    if (app->getPinLocked()) {
-                        std::unique_ptr<LockPhoneData> data =
-                            std::make_unique<LockPhoneData>(LockPhoneData::Request::Pin);
-                        // if there was no application on to before closing proceed normally to pin protection
-                        if (lockTimeoutApplilcation.empty()) {
-                            application->switchWindow(app::window::name::desktop_pin_lock, std::move(data));
-                            return true;
-                        }
-                        else {
-                            data->setPrevApplication(lockTimeoutApplilcation);
-                            lockTimeoutApplilcation = "";
-                            application->switchWindow(app::window::name::desktop_pin_lock, std::move(data));
-                            return true;
-                        }
-                    }
-                    else {
-
-                        // if phone was locked by user show unlocked main window
-                        if (lockTimeoutApplilcation.empty()) {
-                            app->setScreenLocked(false);
-                            setVisibleState();
-                            application->refreshWindow(RefreshModes::GUI_REFRESH_FAST);
-                            return true;
-                        }
-                        // if there was application on top when timeout occurred
-                        else {
-                            lockTimeoutApplilcation = "";
-                            sapm::ApplicationManager::messageSwitchPreviousApplication(application);
-                        }
-                    }
-                }
-                // not locked && not unlock -> show prompt
-                if ((app->getPinLocked() == false) && ((!inputEvent.is(KeyCode::KEY_ENTER)) ||
-                                                       (inputEvent.is(KeyCode::KEY_ENTER) && enter_cache.cached()))) {
-                    if (inputEvent.is(KeyCode::KEY_ENTER) && enter_cache.cached()) {
-                        enter_cache.clear();
-                    }
-                    application->switchWindow(app::window::name::desktop_pin_lock,
-                                              std::make_unique<LockPhoneData>(LockPhoneData::Request::NoPin));
-                    return true;
-                }
-                if (!inputEvent.is(KeyCode::KEY_ENTER)) { // TODO TUTAJ BREAKPOINT
-                    application->switchWindow(app::window::name::desktop_pin_lock,
-                                              std::make_unique<LockPhoneData>(LockPhoneData::Request::ShowPrompt));
-                    return true;
-                }
-                if (enter_cache.storeEnter(inputEvent)) {}
+        if (inputEvent.isLongPress()) {
+            return processLongPressEvent(inputEvent);
+        }
+        else if (inputEvent.isShortPress()) {
+            if (app->lock.isLocked()) {
+                return processShortPressEventOnLocked(inputEvent);
             }
-            // screen is unlocked
             else {
-                // pressing enter moves user to menu screen
-                if (inputEvent.keyCode == KeyCode::KEY_ENTER) {
-                    application->switchWindow("MenuWindow");
-                    return true;
-                }
-                // if numeric key was pressed record that key and send it to call application
-                else if (code != 0) {
-                    return app::prepareCall(application, std::string(1, static_cast<char>(code)));
-                }
+                return processShortPressEventOnUnlocked(inputEvent);
             }
         }
-        else if (inputEvent.isLongPress()) {
-            // long press of # locks screen if it was unlocked
-            if ((inputEvent.keyCode == KeyCode::KEY_PND) && (app->getScreenLocked() == false)) {
-                app->setScreenLocked(true);
-                setVisibleState();
-                app->refreshWindow(RefreshModes::GUI_REFRESH_FAST);
-                app->setSuspendFlag(true);
-                return true;
-            }
-            // long press of right function button move user to power off window
-            else if (inputEvent.keyCode == KeyCode::KEY_RF) {
-                application->switchWindow("PowerOffWindow");
-                return true;
-            }
-            // long press of '0' key is translated to '+'
-            else if (inputEvent.keyCode == KeyCode::KEY_0) {
-                return app::prepareCall(application, "+");
-            }
-        }
-
-        // check if any of the lower inheritance onInput methods catch the event
         return AppWindow::onInput(inputEvent);
     }