+++ changelog.md
@@ -3,9 +3,11 @@
 ## [Current release]
 
 ### Added
-
+* `[desktop][gui]` Added SIM and PUK lock-related windows
+* `[GUI]` Added rich text parsing for full text styling needs
 * `[calendar]` Add notifications from database handling.
 * `[cellular]``[call app]` Added DTMF tone generation 
+* `[settings]` Add basic display and keypad options window
 * `[audio]` Enabled sounds priorities
 
 ### Changed
@@ -13,7 +15,6 @@
 
 
 ### Fixed
-
 
 
 ### Other
@@ -432,3 +433,4 @@
 
 * `[test]` Improved modem testing with modem shutdown detection.
 * `[build]` Add support for auto-releases.
+